Game pauses when player is out of bounds and resumes once the hips begin to track again
Chose to use the hip positioning as it plays a key role in the creation and manipulation of the board through the hip plane.
The hip positioning old X coordinate is tracked and compared to the current X coordinate, as the is changing the positioning almost every frame to ~thousandths place we can confidently use this to see if we have a repeating coordinate which is an indication of it no longer being tracked.